--- src/interfaces/Collidable.h
@@ -1,17 +1,0 @@
-//
-// Created by Jonathan Moallem on 29/1/21.
-//
-
-#ifndef A_DARK_DISCOMFORT_COLLIDABLE_H
-#define A_DARK_DISCOMFORT_COLLIDABLE_H
-
-class Collidable
-{
-public:
-
-    // TODO use this?
-    virtual Entity* GetBounds() = 0;
-
-};
-
-#endif //A_DARK_DISCOMFORT_COLLIDABLE_H
